refactor: update fork and unit tests of NTG for core pool
